Create tileGenerator.py
Main file. All in one place. Requires PIL and NUMPY to be installed. Python version 3.3 was used to create this document. Generate colors method currently doesn't exist as a new one needs to be created since the old version is very hacky and needs to be generalized for dealing with layers. 