defined values, bug fixes, none checks
Fix - The transition row is a 3 dimensional array and it's element is an
array that holds the chance on being at that position in the first
position and the color in the second. Fuck this is wrong.
BUG - 4th dimenstion should be an array of arrays. each position should
be an array that holds a color and a chance of being that color
IDEA - Just fucking make a list of active transitions instead of all of
them. Stupid!