transition system upgrade
IDEA: Fix transition system to a more efficient way of obtaining
direction.
Getting close to image generation again in this version.
Many changes including a transition rework, and adding the start of the
generation system. Might want to look into how to make getting random
numbers to be gathered faster, maybe in batches? Maybe it isn't even a
bottleneck like it was in other languages. GOODNIGHT!